Use field specific types where possible for Job responses (#15)
The bulk package defines types and constants for the valid values
of some job info/option fields. This commit updates the Response
struct to use those where possible rather than the plain string
type. This removes the need to type cast during normal operations
(ie, when checking the state of a Job).